+++ lib/src/core/snackbar_config.dart
@@ -10,6 +10,7 @@ class SmartSnackConfig {
   final Color? backgroundColor;
   final Color? textColor;
   final IconData? icon;
+  final Widget? leading;
 
   const SmartSnackConfig({
     this.duration = const Duration(seconds: 3),
@@ -20,6 +21,7 @@ class SmartSnackConfig {
     this.backgroundColor,
     this.textColor,
     this.icon,
+    this.leading,
   });
 
   factory SmartSnackConfig.forType(SnackType type) {