+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+* Added support for custom leading widgets (images, etc.) via the `leading` parameter.
+* Improved `SmartSnackConfig` to support custom leading widgets.
+
 ## 1.0.0
 
 * Initial release of smart_snackbars.